modified type system. use types from inttypes.h

+++ src/libgtmatch/sarr-def.h
@@ -13,7 +13,7 @@
 
 typedef struct
 {
-  unsigned long numofdbsequences;
+  unsigned int numofdbsequences;
   StrArray *filenametab;
   PairSeqpos *filelengthtab;
   unsigned int prefixlength;
